v6.0-serialize: refactor SigmaDslSpecification towards LanguageSpecifications
